requirements.txt
angler
numpy>=1.19.2
torchvision>=0.9.0.dev20210130
tqdm>=4.56.0
setuptools>=52.0.0
torch>=1.8.0
matplotlib>=3.3.4

tensorflow-gpu>=2.5.0
svglib>=1.1.0
scipy>=1.5.4
scikit-learn>=0.24.1
torchsummary>=1.5.1
pyyaml>=5.1.1
mlflow
